% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize_nosignal}
\alias{categorize_nosignal}
\title{categorize_nosignal. This function is written for determining if the data is no-signal or not}
\usage{
categorize_nosignal(parameterVectorLinear,
  threshold_line_slope_parameter = 0.01, threshold_intensity_interval = 0.1)
}
\arguments{
\item{parameterVectorLinear}{is the output of lineFitFunction}

\item{threshold_line_slope_parameter}{minimum for line slope (Default is 0.01)}

\item{threshold_intensity_interval}{minimum for intensity range (Default is 0.1)}
}
\value{
Function simply returns one of the three text outputs. "no_signal", "NOT no_signal"
}
\description{
The decision of no_signal or NOT no_signal does not depend on sigmoidal or double_sigmoidal fits. Most of the time a high percent of the data is composed of no signal cells. So labelling no_signal cells without doing sigmoidal and double sigmoidal fit make all process faster. This function is designed to do that.
}
\examples{
# Example 1 with double sigmoidal data
# Initial Command to Reset the System
rm(list = ls())
if (is.integer(dev.list())){dev.off()}
cat("\\014")

time=seq(3,24,0.1)

#intensity with Noise
noise_parameter=0.2
intensity_noise=runif(n = length(time),min = 0,max = 1)*noise_parameter
intensity=doublesigmoidalFitFormula(time,
                                   finalAsymptoteIntensity=.3,
                                   maximum=4,
                                   slope1=1,
                                   midPoint1=7,
                                   slope2=1,
                                   midPointDistance=8)
intensity=intensity+intensity_noise

dataInput=data.frame(intensity=intensity,time=time)
dataOutput = normalizeData(dataInput,dataInputName="batch_01_21_2016_samp007623")
dataInput2=dataOutput


# Do the line fit
parameterVectorLinear=fitFunction(dataInput=dataInput2,
                                 model="linear",
                                 n_runs_min=20,
                                 n_runs_max=500,
                                 showDetails=FALSE)

isThis_nosignal=categorize_nosignal(parameterVectorLinear=parameterVectorLinear)



# Example 2 with no_signal data
# Initial Command to Reset the System
rm(list = ls())
if (is.integer(dev.list())){dev.off()}
cat("\\014")

time=seq(3,24,0.1)

#intensity with Noise
noise_parameter=0.05
intensity_noise=runif(n = length(time),min = 0,max = 1)*noise_parameter*2e-04
intensity=doublesigmoidalFitFormula(time,
                                   finalAsymptoteIntensity=.3,
                                   maximum=2e-04,
                                   slope1=1,
                                   midPoint1=7,
                                   slope2=1,
                                   midPointDistance=8)
intensity=intensity+intensity_noise

dataInput=data.frame(intensity=intensity,time=time)
dataOutput = normalizeData(dataInput,dataInputName="batch_01_21_2016_samp007623")
dataInput2=dataOutput


# Do the line fit
parameterVectorLinear=fitFunction(dataInput=dataInput2,
                                 model="linear",
                                 n_runs_min=20,
                                 n_runs_max=500,
                                 showDetails=FALSE)

isThis_nosignal=categorize_nosignal(parameterVectorLinear=parameterVectorLinear)

}

