% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/counterFunction.R
\name{counterFunction}
\alias{counterFunction}
\title{Counter function}
\usage{
counterFunction(dataInput, model, n_runs_min, n_runs_max, ...)
}
\arguments{
\item{model}{type of fit function that will be used. Can be "linear", "sigmoidal", "double_sigmoidal", "test"}

\item{n_runs_min}{number of minimum runs that the algorithm can run}

\item{n_runs_max}{number of maximum runs that the algorithm can run}

\item{data}{Normalized input data that will be fitted transferred into related functions}
}
\value{
The function returns the parameters related with fitted curve to input data
}
\description{
Counter function
}
\details{
the algorithm calls the fitting algorithms. to make the fits with random initial parameters. This multiple runs are necessary to avoid local minimums that LM fits can stuck. Fitting algorithms can either gives a fit with related parameters and isThisaFit=TRUE value or just give isThisaFit=FALSE corresponding to not a fit. n_runs_min represent minimum number of fits that are necessary to give a result, n_runs_max limits the number of runs (successful or unsuccessful) that the it algorithm can run
}
\examples{
# Example 1 (test function without normalization)
# data sent to algorithm directly as data frame
# a- Generate data
time = seq(3,48,0.5)
intensity=runif(length(time), 3.0, 7.5)
dataInput = data.frame(time,intensity)
# b- generate "random Parameter" for model "test"
randomParameter=0.7 # it should be a parameter between 0 and 1
# c- use the function "test"
parameterOutput=counterFunction(data=dataInput,model="test",n_runs_min=5,n_runs_max=15)

# Example 2 (test function with normalization)
# data sent to algorithm after normalization
# a- Generate data
time = seq(3,48,0.5)
intensity=runif(length(time), 3.0, 7.5)
dataInput = data.frame(time,intensity)
# b- normalize data
dataOutput = normalizeData(dataInput)
# c- generate "random Parameter" for model "test"
randomParameter=0.7 # it should be a parameter between 0 and 1
# d- use the function "test"
dataInput2=dataOutput
parameterOutput=counterFunction(data=dataInput2,
                                model="test",
                                n_runs_min=5,
                                n_runs_max=15)

# Example 3 (linear function without normalization)
# data sent to algorithm directly as data frame
# a- Generate data
time = seq(3,48,0.5)
intensity=runif(length(time), 3.0, 7.5)
dataInput = data.frame(time,intensity)
# b- use the function "linear"
parameterOutput=counterFunction(data=dataInput,
                                model="linear",
                                n_runs_min=5,
                                n_runs_max=15)

# Example 4 (linear function with normalization)
# data sent to algorithm after normalization
# a- Generate data
time = seq(3,48,0.5)
intensity=runif(length(time), 3.0, 7.5)
dataInput = data.frame(time,intensity)
# b- normalize data
dataOutput = normalizeData(dataInput)
# c- use the function "linear"
dataInput2=dataOutput
parameterOutput=counterFunction(data=dataInput2,
                                model="linear",
                                n_runs_min=5,
                                n_runs_max=15)
}

