% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{exampleFitFunction}
\alias{exampleFitFunction}
\title{exampleFitFunction}
\usage{
exampleFitFunction(randomParameter)
}
\arguments{
\item{randomParameter}{This parameter defines the probability that the exampleFitFunction returns TRUE values for isThisaFit parameter. The aparemeter should be in the interval of 0 and 1}
}
\value{
The function returns TRUE or FALSE for isThisaFit parameter and also residual_Sum_of_Squares parameter that determines the goodness of the fit
}
\description{
exampleFitFunction
}
\details{
This is the exampleFitFunction that generates TRUE values for isThisaFit parameter whit given probability
}
\examples{
# add usage examples here
}

