% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublesigmoidalFitFunctions.R
\name{f_argmax_doublesigmoidal}
\alias{f_argmax_doublesigmoidal}
\title{f_argmax_doublesigmoidal}
\usage{
f_argmax_doublesigmoidal(parameterDf)
}
\arguments{
\item{parameterDf}{is a dataFrame that includes columns with names:

*"dataScalingParameters.timeRatio" that represents maximum time of intensity measurements,

parameters of double sigmoidal model:
 *"slope1_Estimate",
 *"midPoint1_Estimate",
 *"slope2_Estimate",
 *"midPointDistance_Estimate".
 (Note: additional parameters in Df does not cause problems)

finally a column with the used model information model="doublesigmoidal".}
}
\value{
return the x value of the maximum in between time 0 and maximum time measured in sample intensities. In other words the time that double sigmoidal function reaches its maximum
}
\description{
This function is designed to compansate the small distribuncy in numerical parameters in double sigmoidal model numerically. Function is called from numericalReCalculation function.
}
\examples{
# Related example

# Initial Command to Reset the System
rm(list = ls())
if (is.integer(dev.list())){dev.off()}
cat("\\014")

# Generate a parameters DF
parameterDf=data.frame(dataScalingParameters.timeRatio=24,
                      slope1_Estimate=1,
                      midPoint1_Estimate=7,
                      slope2_Estimate=1,
                      midPointDistance_Estimate=8,
                      model="doublesigmoidal")

# Find the x value of the maximum for this model
# with in the range of [0, dataScalingParameters.timeRatio]
maximum_x=f_argmax_doublesigmoidal(parameterDf)

}

