% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublesigmoidalFitFunctions.R
\name{f_mid1_doublesigmoidal}
\alias{f_mid1_doublesigmoidal}
\title{f_mid1_doublesigmoidal}
\usage{
f_mid1_doublesigmoidal(parameterVector)
}
\value{
return the x values of midpoint1, by assuming the y of midpoint1 is the half of the maximum value it reaches
}

