% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublesigmoidalFitFunctions.R
\name{f_slope_doublesigmoidal}
\alias{f_slope_doublesigmoidal}
\title{f_slope_doublesigmoidal}
\usage{
f_slope_doublesigmoidal(x, parameterDf, timeStep = 1e-05)
}
\arguments{
\item{x}{as a single time point or a sequence of time points that one wants to calculate deritive.}

\item{parameterDf}{is a dataFrame that includes columns with names:

*"dataScalingParameters.timeRatio" that represents maximum time of intensity measurements,

parameters of double sigmoidal model:
 *"finalAsymptoteIntensity_Estimate",
 *"maximum_Estimate",
 *"slope1_Estimate",
 *"midPoint1_Estimate",
 *"slope2_Estimate",
 *"midPointDistance_Estimate".

finally a column with the used model information model="doublesigmoidal".}

\item{timeStep}{is the time step for the derivative calculation increment "h". Default value is 0.00001}
}
\value{
return the numerical slope of double sigmoidal for the given x value
}
\description{
The function calculates numerical slope of double sigmoidal function with given parameters by using 5 points derivative. It is designed to compansate the small distribuncy in numerical parameters in double sigmoidal model numerically. Function is called from numericalReCalculation function.
}
\examples{
# Related examples

# Example 1
# Initial Command to Reset the System
rm(list = ls())
if (is.integer(dev.list())){dev.off()}
cat("\\014")

# Generate a parameters DF
parameterDataframe=data.frame(dataScalingParameters.timeRatio=24,
                             finalAsymptoteIntensity_Estimate=.3,
                             maximum_Estimate=4,
                             slope1_Estimate=1,
                             midPoint1_Estimate=7,
                             slope2_Estimate=1,
                             midPointDistance_Estimate=8,
                             model="doublesigmoidal")

# Find the x value of the maximum for this model
# with in the range of [0, dataScalingParameters.timeRatio]
slope_Estimate=f_slope_doublesigmoidal(x=6.5,
                                       parameterDf=parameterDataframe,
                                       timeStep=0.00001)
# Example 2
# x might be vector as well
# Initial Command to Reset the System
rm(list = ls())
if (is.integer(dev.list())){dev.off()}
cat("\\014")

# Generate a parameters DF
parameterDataframe=data.frame(dataScalingParameters.timeRatio=24,
                             finalAsymptoteIntensity_Estimate=.3,
                             maximum_Estimate=4,
                             slope1_Estimate=1,
                             midPoint1_Estimate=7,
                             slope2_Estimate=1,
                             midPointDistance_Estimate=8,
                             model="doublesigmoidal")

# Find the x value of the maximum for this model
# with in the range of [0, dataScalingParameters.timeRatio]
slope_Estimate=f_slope_doublesigmoidal(x=seq(-6.5,50,0.5),
                                       parameterDf=parameterDataframe,
                                       timeStep=0.00001)
# Note that some points in the x sequence are out of the range of
# time = [0, dataScalingParameters.timeRatio]

}

