% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fGeneral_00.R
\name{fGeneral_00}
\alias{fGeneral_00}
\title{Title}
\usage{
fGeneral_00(x, slope, intersection)
}
\arguments{
\item{x}{}

\item{slope}{}

\item{intersection}{}
}
\value{
Describe here what the function returns.
}
\description{
Title
}
\examples{
# add usage examples here
}

