% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fGeneral_00_Numeric.R
\name{fGeneral_00_Numeric}
\alias{fGeneral_00_Numeric}
\title{Title}
\usage{
fGeneral_00_Numeric(newColumnName, timeInput, functionInput)
}
\arguments{
\item{newColumnName}{}

\item{timeInput}{}

\item{functionInput}{}
}
\value{
Describe here what the function returns.
}
\description{
Title
}
\examples{
# add usage examples here
}

