% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterFunction.R
\name{filterFunction}
\alias{filterFunction}
\title{Title}
\usage{
filterFunction(filterNo, virusData)
}
\arguments{
\item{filterNo}{}

\item{virusData}{}
}
\value{
Describe here what the function returns.
}
\description{
Title
}
\examples{
# add usage examples here
}

