% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_main.R, R/fit_main2.R
\name{fit_main}
\alias{fit_main}
\title{Title}
\usage{
fit_main(data, fitFunction, startList, lowerBounds, upperBounds, min_Factor,
  n_iterations, n_runs_max, n_runs_min)

fit_main(data, fitFunction, startList, lowerBounds, upperBounds, min_Factor,
  n_iterations, n_runs_max, n_runs_min)
}
\arguments{
\item{data}{}

\item{fitFunction}{}

\item{startList}{}

\item{lowerBounds}{}

\item{upperBounds}{}

\item{min_Factor}{}

\item{n_iterations}{}

\item{n_runs_max}{}

\item{n_runs_min}{}

\item{data}{Normalized input data that will be fitted}

\item{fitFunction}{type of fit function that will be used. Can be "linear", "sigmoidal", "double-sigmoidal"}
}
\value{
Describe here what the function returns.

The function returns the parameters related with fitted curve to input data
}
\description{
Title

Title
}
\examples{
# add usage examples here
# add usage examples here
}

