% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_noSol.R
\name{fit_noSol}
\alias{fit_noSol}
\title{Title}
\usage{
fit_noSol(startList)
}
\arguments{
\item{startList}{}
}
\value{
Describe here what the function returns.
}
\description{
Title
}
\examples{
# add usage examples here
}

