% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_S1.R
\name{fit_S1}
\alias{fit_S1}
\title{Title}
\usage{
fit_S1(data, fitFunction, startList, lowerBounds, upperBounds, min_Factor,
  n_iterations)
}
\arguments{
\item{data}{}

\item{fitFunction}{}

\item{startList}{}

\item{lowerBounds}{}

\item{upperBounds}{}

\item{min_Factor}{}

\item{n_iterations}{}
}
\value{
Describe here what the function returns.
}
\description{
Title
}
\examples{
# add usage examples here
}

