% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublesigmoidalFitFunctions.R
\name{fslope_doublesigmoidal}
\alias{fslope_doublesigmoidal}
\title{Numerical Slope of Double Sigmoidal}
\usage{
fslope_doublesigmoidal(x, parameterVector, timeStep = 1e-05)
}
\value{
return the numerical slope of double sigmoidal for the given x value
}

