% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lineFitFunctions.R
\name{lineFitFunction}
\alias{lineFitFunction}
\title{lineFitFunction}
\usage{
lineFitFunction(dataInput, tryCounter, startList = list(slope = 0,
  intersection = 1), lowerBounds = c(-100, -1000), upperBounds = c(100,
  1000), min_Factor = 1/2^20, n_iterations = 500)
}
\arguments{
\item{dataInput}{dataInput should be a data frame composed of two columns. One is for time other is for intensity. Or it should be a normalized data generated by normalizeData function}

\item{tryCounter}{is the input that represent the number of the call of this function for the same situation}

\item{startList}{The initial set of parameters that algorithm tries for the fit, where the parameters are slope and intercetion.}

\item{lowerBounds}{The lower bouns for the randomly generated start parameters, where the parameters are slope and intercetion.}

\item{upperBounds}{The lower bouns for the randomly generated start parameters, where the parameters are slope and intercetion.}

\item{min_Factor}{define minimum step size in the iterations}

\item{n_iterations}{define maximum number of iterations for a run}
}
\value{
Function returns fitted parameters for lineFit. The slope, intersection and parameters related with quality of the fit
}
\description{
lineFitFunction
}
\examples{
# related examples
intensity=c(0.9947917, 0.7395833, 0.4010417, 0.8020833, 0.5364583, 0.6718750, 0.6770833,
      0.6145833, 0.6041667, 0.2708333, 1.0000000, 0.4687500, 0.3385417, 0.5364583,
      0.4687500, 0.6041667, 0.6718750, 0.3333333, 0.4010417, 0.7500000, 0.2031250,
      0.4114583, 0.4010417, 0.8802083, 0.6041667, 0.7500000, 0.4010417, 0.8072917,
      0.4687500, 0.3333333, 0.2708333, 0.0000000, 0.4010417, 0.7343750, 0.4062500,
      0.6718750, 0.6666667, 0.3385417, 0.8020833, 0.5468750, 0.6093750, 0.6093750, 0.6770833)
time=seq(3,24,0.5)
dataInput=data.frame(intensity=intensity,time=time)
parameterVector<-lineFitFunction(dataInput,tryCounter=2)
}

