% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterCalculation.R
\name{parameterCalculation}
\alias{parameterCalculation}
\title{useful paramter calculation with help of fits}
\usage{
parameterCalculation(parameterVector, stepSize = 1e-05)
}
\arguments{
\item{parameterVector}{output of fitFunction or data frame that gives the variables related with double sigmoidal fit.}

\item{stepSize}{step size used by the fitting algorithm.}
}
\value{
Returns the expanded paramter vectors; that includes externally meaningful values as well as internal parameters generated from fit algorithms.
}
\description{
Generates useful values for external use, with the help of parameterVector's of the fits.
}
\examples{
time=seq(3,24,0.1)

#simulate intensity data with noise
noise_parameter=0.2
intensity_noise = stats::runif(n = length(time),min = 0,max = 1)*noise_parameter
intensity = sicegar::doublesigmoidalFitFormula(time,
                                            finalAsymptoteIntensityRatio=.3,
                                            maximum=4,
                                            slope1Param=1,
                                            midPoint1Param=7,
                                            slope2Param=1,
                                            midPointDistanceParam=8)
intensity = intensity+intensity_noise

dataInput=data.frame(intensity=intensity,time=time)
normalizedInput = sicegar::normalizeData(dataInput)
parameterVector = sicegar::multipleFitFunction(dataInput=normalizedInput,
                                            dataInputName="sample01",
                                            model="doublesigmoidal",
                                            n_runs_min=20,
                                            n_runs_max=500,
                                            showDetails=FALSE)

if(parameterVector$isThisaFit)
{parameterVector <- sicegar::parameterCalculation(parameterVector)}
print(t(parameterVector))

}

