% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{pre_categorize}
\alias{pre_categorize}
\title{Checks for signal in the data.}
\usage{
pre_categorize(normalizedInput, threshold_intensity_range = 0.1,
  threshold_minimum_for_intensity_maximum = 0.3)
}
\arguments{
\item{normalizedInput}{is the output of sicegar::normalizeData function.}

\item{threshold_intensity_range}{minimum for intensity range (Default is 0.1).}

\item{threshold_minimum_for_intensity_maximum}{minimum allowed value for intensity maximum}
}
\value{
Function returns decision vector that at the end says "no signal" or not "no signal".
}
\description{
Checks if signal is present in the data. Often a high percentage of high through-put data does not contain a signal. Checking if data does not contain signal before doing a sigmoidal or double sigmoidal fit can make analysis of data from high through-put experiments much faster.
}
\examples{
# Example 1 with double sigmoidal data

time=seq(3,24,0.1)

#simulate intensity data and add noise
noise_parameter=0.2
intensity_noise=runif(n = length(time),min = 0,max = 1)*noise_parameter
intensity=sicegar::doublesigmoidalFitFormula(time,
                                   finalAsymptoteIntensityRatio=.3,
                                   maximum=4,
                                   slope1Param=1,
                                   midPoint1Param=7,
                                   slope2Param=1,
                                   midPointDistanceParam=8)
intensity=intensity+intensity_noise

dataInput=data.frame(intensity=intensity,time=time)
normalizedInput = sicegar::normalizeData(dataInput,dataInputName="batch_01_21_2016_samp007623")
isThis_nosignal = sicegar::pre_categorize(normalizedInput = normalizedInput)



# Example 2 with no_signal data

time=seq(3,24,0.1)

#simulate intensity data and add noise
noise_parameter=0.05
intensity_noise=runif(n = length(time),min = 0,max = 1)*noise_parameter*2e-04
intensity=sicegar::doublesigmoidalFitFormula(time,
                                   finalAsymptoteIntensityRatio=.3,
                                   maximum=2e-04,
                                   slope1Param=1,
                                   midPoint1Param=7,
                                   slope2Param=1,
                                   midPointDistanceParam=8)
intensity=intensity+intensity_noise

dataInput=data.frame(intensity=intensity,time=time)
normalizedInput = sicegar::normalizeData(dataInput,dataInputName="batch_01_21_2016_samp007623")
isThis_nosignal = sicegar::pre_categorize(normalizedInput = normalizedInput)

}

