% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFigures.R
\name{printInfectionCurves}
\alias{printInfectionCurves}
\title{printInfectionCurves}
\usage{
printInfectionCurves(dataInput, sigmoidalFitVector = NULL,
  doubleSigmoidalFitVector = NULL, showParameterRelatedLines = FALSE)
}
\arguments{
\item{dataInput}{dataInput should be a data frame composed of two columns. One is for time other is for intensity. Or it should be a normalized data generated by normalizeData function}

\item{sigmoidalFitVector}{is the output of sigmoidalFitFunction. Default is NULL}

\item{doubleSigmoidalFitVector}{is the output of double sigmoidal fit function. Default is NULL}

\item{showParameterRelatedLines}{show parameter related lines on the curves}
}
\value{
The function returns infection curve figures by using ggplot
}
\description{
The aim of the function is to generate figures by using ggplot that shows input data and fit curves.
}
\examples{
# Example 1 with double sigmoidal data
# Initial Command to Reset the System
rm(list = ls())
if (is.integer(dev.list())){dev.off()}
cat("\\014")

time=seq(3,24,0.1)

#intensity with Noise
noise_parameter=0.2
intensity_noise=runif(n = length(time),min = 0,max = 1)*noise_parameter
intensity=doublesigmoidalFitFormula(time,
                                   finalAsymptoteIntensity=.3,
                                   maximum=4,
                                   slope1=1,
                                   midPoint1=7,
                                   slope2=1,
                                   midPointDistance=8)
intensity=intensity+intensity_noise

dataInput=data.frame(intensity=intensity,time=time)
dataOutput = normalizeData(dataInput,dataInputName="batch_01_21_2016_samp007623")
dataInput2=dataOutput

# Do the sigmoidal fit
parameterVectorSigmoidal=fitFunction(dataInput=dataInput2,
                                    model="sigmoidal",
                                    n_runs_min=20,
                                    n_runs_max=500,
                                    showDetails=FALSE)

# Do the double sigmoidal fit
parameterVectorDoubleSigmoidal=fitFunction(dataInput=dataInput2,
                                          model="doublesigmoidal",
                                          n_runs_min=20,
                                          n_runs_max=500,
                                          showDetails=FALSE)

parameterVectorDoubleSigmoidal = numericalReCalculation(parameterVectorDoubleSigmoidal,
                                                       stepSize=0.00001)
fig01=printInfectionCurves(dataInput=dataInput2)
print(fig01)

fig02=printInfectionCurves(dataInput=dataInput2,
                          sigmoidalFitVector=parameterVectorSigmoidal)
print(fig02)

fig03=printInfectionCurves(dataInput=dataInput2,
                          doubleSigmoidalFitVector=parameterVectorDoubleSigmoidal)
print(fig03)

fig04=printInfectionCurves(dataInput=dataInput2,
                          sigmoidalFitVector=parameterVectorSigmoidal,
                          doubleSigmoidalFitVector=parameterVectorDoubleSigmoidal)
print(fig04)

fig05=printInfectionCurves(dataInput=dataInput2,
                          doubleSigmoidalFitVector=parameterVectorDoubleSigmoidal,
                          showParameterRelatedLines=TRUE)
print(fig05)

fig06=printInfectionCurves(dataInput=dataInput2,
                          sigmoidalFitVector=parameterVectorSigmoidal,
                          showParameterRelatedLines=TRUE)
print(fig06)

}

