% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/renameFun.R
\name{renameFun}
\alias{renameFun}
\title{Title}
\usage{
renameFun(dat, newnames, oldnames)
}
\arguments{
\item{dat}{}

\item{newnames}{}

\item{oldnames}{}
}
\value{
Describe here what the function returns.
}
\description{
Title
}
\examples{
# add usage examples here
}

