% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublesigmoidalFItFunctions_h0.R
\name{renormalizeParameters_h0}
\alias{renormalizeParameters_h0}
\title{Convert normalized min.pack estimated parameters back to raw scale (with h0)}
\usage{
renormalizeParameters_h0(parameterDF, isalist)
}
\arguments{
\item{parameterDF}{Data.frame with columns `*_N_Estimate` and
`dataScalingParameters.*`.}

\item{isalist}{Logical; if `TRUE` input came from a list,
otherwise already raw.}
}
\value{
A one‐row data.frame with `*_Estimate` columns on the raw
  scale (including `h0_Estimate`).
}
\description{
Takes a one‐row parameter data.frame and “unscales” its
time‐normalized and intensity‐normalized columns.
}
\examples{
df <- data.frame(
  model="doublesigmoidal",
  maximum_N_Estimate=1, finalAsymptoteIntensityRatio_N_Estimate = .5,
  slope1Param_N_Estimate=1, midPoint1Param_N_Estimate = .3,
  slope2Param_N_Estimate=1, midPointDistanceParam_N_Estimate = .2,
  h0_N_Estimate = .1,
  dataScalingParameters.timeRange = 100,
  dataScalingParameters.intensityRange = 10,
  dataScalingParameters.intensityMin = 0
)
renormalizeParameters_h0(df, isalist=TRUE)
}
