% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFigures.R
\name{sameSourceDataCheck}
\alias{sameSourceDataCheck}
\title{sameSourceDataCheck}
\usage{
sameSourceDataCheck(dataInput, sigmoidalFitVector, doubleSigmoidalFitVector)
}
\arguments{
\item{dataInput}{dataInput should be a data frame composed of two columns. One is for time other is for intensity. Or it should be a normalized data generated by normalizeData function}

\item{sigmoidalFitVector}{is the output of sigmoidalFitFunction. Default is NULL}

\item{doubleSigmoidalFitVector}{is the output of double sigmoidal fit function. Default is NULL}
}
\value{
The function checks if the data and models came from the same source
}
\description{
The aim of the function is to check if the provied data and provided models came from same source by looking to ".dataInputName" columns of the inputs.
}

