\name{sicegar-package}
\alias{sicegar-package}
\alias{sicegar}
\docType{package}
\title{
\packageTitle{sicegar}
}
\description{
\packageDescription{sicegar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sicegar}
\packageIndices{sicegar}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sicegar}

Maintainer: \packageMaintainer{sicegar}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
