% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizationFunction_h0.R
\name{unnormalizeData_h0}
\alias{unnormalizeData_h0}
\title{Unnormalize Time–Intensity Data}
\usage{
unnormalizeData_h0(dataInput)
}
\arguments{
\item{dataInput}{A list as returned by \code{\link{normalizeData_h0}},
containing \code{timeIntensityData} and \code{dataScalingParameters}.}
}
\value{
A list with component:
  \item{timeIntensityData}{data.frame of original \code{time} and
    \code{intensity} values.}
}
\description{
Reverses the transformation applied by \code{\link{normalizeData_h0}},
restoring raw time and intensity values from scaled data.
}
\examples{
norm <- normalizeData_h0(data.frame(time=0:10, intensity=rnorm(11)))
unnormalizeData_h0(norm)
}
