% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataInputCheck.R
\name{dataCheck}
\alias{dataCheck}
\title{DataCheck Function}
\usage{
dataCheck(data)
}
\arguments{
\item{data}{it is the input data file. In can be either a list that contains a data frame in .$timeIntensityData or a data frame by itself.}
}
\description{
DataCheck Function
}
\details{
The dataCheck checks if the input data is appropirate and covert it into an appropirate form. Data frame should contain two columns named time and intensity for timeData and intensityData. If the data frame is in a list its name in the list should be $timeIntensityData
}
\examples{
# Example 1

# Initial Command to Reset the System
rm(list = ls())
if (is.integer(dev.list())){dev.off()}
cat("\\014")

# generate data frame
time = seq(3,48,0.5)
intensity=runif(length(time), 3.0, 7.5)
dataInput = data.frame(time,intensity)

# Apply dataCheck function
dataOutputVariable = dataCheck(dataInput)

# Example 2

# Initial Command to Reset the System
rm(list = ls())
if (is.integer(dev.list())){dev.off()}
cat("\\014")

# generate data frame
time = seq(3,48,0.5)
intensity=runif(length(time), 3.0, 7.5)
dataInput = data.frame(time,intensity)

# Normalize Data
dataOutput = normalizeData(dataInput)
dataInput2=dataOutput

# Apply dataCheck function?
dataOutputVariable2 = dataCheck(dataInput2)
}

