% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublesigmoidalFitFunctions.R
\name{f_argmax_doublesigmoidal}
\alias{f_argmax_doublesigmoidal}
\title{f_argmax_doublesigmoidal}
\usage{
f_argmax_doublesigmoidal(parameterVector)
}
\value{
return the x value of the maximum
}

