% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublesigmoidalFitFunctions.R
\name{f_mid2_doublesigmoidal}
\alias{f_mid2_doublesigmoidal}
\title{f_mid2_doublesigmoidal}
\usage{
f_mid2_doublesigmoidal(parameterVector)
}
\value{
return the x values of midpoint2, by assuming the y of midpoint2 is the half of the distance between maximum value it reaches and final asymptote
}

