% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericalReCalculation.R
\name{numericalReCalculation}
\alias{numericalReCalculation}
\title{Generates derived parameters from fit parameters of sigmoidal and double sigmoidal fit. Some of them calculated directlt from fit parameters and some of them are calculated from numerical computations from fitline generated from fit parameters.}
\usage{
numericalReCalculation(parameterVector, stepSize = 1e-05)
}
\arguments{
\item{parameterVector}{output of fitFunction or data frame that gives the variables related with double sigmoidal fit.}

\item{stepSize}{step size used by the fitting algorithm.}
}
\value{
Returns the parameters related with fitted curve to input data.
}
\description{
Calls the fitting algorithms and fits the data  with random initial parameters.
}
\examples{
time=seq(3,24,0.1)

#simulate intensity data with noise
noise_parameter=0.2
intensity_noise=stats::runif(n = length(time),min = 0,max = 1)*noise_parameter
intensity=doublesigmoidalFitFormula(time,
                                   finalAsymptoteIntensityRatio=.3,
                                   maximum=4,
                                   slope1Param=1,
                                   midPoint1Param=7,
                                   slope2Param=1,
                                   midPointDistanceParam=8)
intensity=intensity+intensity_noise

dataInput=data.frame(intensity=intensity,time=time)
normalizedInput = normalizeData(dataInput)
parameterVector=fitFunction(dataInput=normalizedInput,
                           dataInputName="batch_01_21_2016_samp007623",
                           model="doublesigmoidal",
                           n_runs_min=20,
                            n_runs_max=500,
                            showDetails=FALSE)

if(parameterVector$isThisaFit)
{parameterVector<-sicegar::numericalReCalculation(parameterVector)}
(parameterVector)

}

