% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureGeneration.R
\name{sameSourceDataCheck}
\alias{sameSourceDataCheck}
\title{Check is data came from the same source.}
\usage{
sameSourceDataCheck(dataInput, sigmoidalFitVector, doubleSigmoidalFitVector)
}
\arguments{
\item{dataInput}{a data frame composed of two columns. One is for time and the other is for intensity. Should be normalized data generated by normalizeData.}

\item{sigmoidalFitVector}{is the output of sigmoidalFitFunction. Default is NULL.}

\item{doubleSigmoidalFitVector}{is the output of double sigmoidal fit function. Default is NULL.}
}
\value{
Returns TRUE if models can from same source, FALSE otherwise.
}
\description{
Checks if the provided data and models came from same source by looking to ".dataInputName" columns of the inputs.
}
