% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoidalFitFunctions.R
\name{sigmoidalFitFunction}
\alias{sigmoidalFitFunction}
\title{sigmoidalFitFunction}
\usage{
sigmoidalFitFunction(dataInput, tryCounter, startList = list(maximum = 1,
  slope = 36, midPoint = 0.3333333), lowerBounds = c(maximum = 0.3, slope =
  1e-05, midPoint = 0.3125 - 0.8333333), upperBounds = c(maximum = 1.5, slope
  = 180, midPoint = 0.3125 + 0.8333333), min_Factor = 1/2^20,
  n_iterations = 1000)
}
\arguments{
\item{dataInput}{dataInput should be a data frame composed of two columns. One is for time other is for intensity. Or it should be a normalized data generated by normalizeData function}

\item{tryCounter}{is the input that represent the number of the call of this function for the same situation.}

\item{startList}{The initial set of parameters that algorithm tries for the fit. Where the parameters are 'maximumValue' that represents the maximum value that the function that can take,  'slopeValue' that represents the slope in normalized y axis, 'midPointValue' that represents midpoint.}

\item{lowerBounds}{The lower bouns for the randomly generated start parameters. The elements of the vector are  'maximumValue' that represents the maximum value that the function that can take,  'slopeValue' that represents the slope in normalized y axis, 'midPointValue' that represents midpoint}

\item{upperBounds}{The lower bouns for the randomly generated start parameters. The elements of the vector are  'maximumValue' that represents the maximum value that the function that can take,  'slopeValue' that represents the slope in normalized y axis, 'midPointValue' that represents midpoint}

\item{min_Factor}{define minimum step size in the iterations}

\item{n_iterations}{define maximum number of iterations for a run}
}
\value{
Function returns fitted parameters for lineFit. The slope, intersection and parameters related with quality of the fit
}
\description{
the function fits a sigmoidal curve to given data by using likelihood maximization (LM) and gives the parameters (maximum, slope and midpoint) describing the sigmoidal fit as output. It also provides information about goodness of the fit as AIC, BIC values, residual_Sum_of_Squares and log_likelihood
}
\examples{
# related examples

# Initial Command to Reset the System
rm(list = ls())
if (is.integer(dev.list())){dev.off()}
cat("\\014")

time=seq(3,24,0.5)

#intensity with Noise
noise_parameter=0.1
intensity_noise=stats::runif(n = length(time),min = 0,max = 1)*noise_parameter
intensity=sigmoidalFitFormula(time, maximum=4, slope=1, midPoint=8)
intensity=intensity+intensity_noise

dataInput=data.frame(intensity=intensity,time=time)
dataOutput = normalizeData(dataInput)
dataInput2=dataOutput
parameterVector<-sigmoidalFitFunction(dataInput2,tryCounter=2)

#Check the results
if(parameterVector$isThisaFit){
intensityTheoretical=sigmoidalFitFormula(time,
                                        maximum=parameterVector$maximum_Estimate,
                                        slope=parameterVector$slope_Estimate,
                                        midPoint=parameterVector$midPoint_Estimate)

comparisonData=cbind(dataInput,intensityTheoretical)
require(ggplot2)
ggplot(comparisonData)+
 geom_point(aes(x=time, y=intensity))+
 geom_line(aes(x=time,y=intensityTheoretical))+
 expand_limits(x = 0, y = 0)}

if(!parameterVector$isThisaFit){print(parameterVector)}

}

